requirements.txt
requests>=2.25.0
bip-utils>=2.7.0
pynacl>=1.5.0